Improved accuracy of chi2 probability calculations

+++ include/presto.h
@@ -512,11 +512,11 @@ double equivalent_gaussian_sigma(double logp);
 /* Return the approximate significance in Gaussian sigmas */
 /* corresponding to a natural log probability logp        */
 
-double chi2_logp(double chi2, int dof);
+double chi2_logp(double chi2, double dof);
 /* Return the natural log probability corresponding to a chi^2 value */
 /* of chi2 given dof degrees of freedom. */
 
-double chi2_sigma(double chi2, int dof);
+double chi2_sigma(double chi2, double dof);
 /* Return the approximate significance in Gaussian sigmas        */
 /* sigmas of a chi^2 value of chi2 given dof degrees of freedom. */
